@@ -13,25 +13,25 @@
     "node": "^18.13.0 || ^20.9.0"
   },
   "dependencies": {
-    "@angular/animations": "^17.2.4",
+    "@angular/animations": "^18.0.3",
     "@angular/cdk": "^17.2.2",
-    "@angular/common": "^17.2.4",
-    "@angular/compiler": "^17.2.4",
-    "@angular/core": "^17.2.4",
-    "@angular/forms": "^17.2.4",
+    "@angular/common": "^18.0.3",
+    "@angular/compiler": "^18.0.3",
+    "@angular/core": "^18.0.3",
+    "@angular/forms": "^18.0.3",
     "@angular/material": "^17.2.2",
     "@angular/material-experimental": "^17.2.2",
-    "@angular/platform-browser": "^17.2.4",
-    "@angular/platform-browser-dynamic": "^17.2.4",
-    "@angular/router": "^17.2.4",
+    "@angular/platform-browser": "^18.0.3",
+    "@angular/platform-browser-dynamic": "^18.0.3",
+    "@angular/router": "^18.0.3",
     "rxjs": "~7.8.0",
     "tslib": "^2.3.0",
     "zone.js": "~0.14.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.2.3",
-    "@angular/cli": "~17.2.3",
-    "@angular/compiler-cli": "^17.2.4",
+    "@angular-devkit/build-angular": "^18.0.4",
+    "@angular/cli": "~18.0.4",
+    "@angular/compiler-cli": "^18.0.3",
     "@types/jasmine": "~4.3.0",
     "jasmine-core": "~4.6.0",
     "karma": "~6.4.0",
@@ -39,6 +39,6 @@
     "karma-coverage": "~2.2.0",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.0.0",
-    "typescript": "~5.2.2"
+    "typescript": "~5.4.5"
   }
 }
